Changes done for delete and styling

+++ app.js
@@ -1,46 +1,43 @@
 const todoTask = document.querySelector("form");
-const clearItems = document.querySelector(".clear")
+const clearItems = document.querySelector(".clear");
 const checkClick = document.querySelector("ul");
 
-checkClick.addEventListener('click', (e) =>{
-console.log(e.target.name);
-  if(e.target.name == 'checkButton')
-     checkTodo(e);
-  if(e.target.name == 'deleteButton')
-    deleteTodo(e);
-
-})
+checkClick.addEventListener("click", (e) => {
+  console.log(e.target.name);
+  if (e.target.name == "checkButton") checkTodo(e);
+  if (e.target.name == "deleteButton") deleteTodo(e);
+});
 
 const checkTodo = (handleEvent) => {
   let item = handleEvent.target.parentNode;
   console.log(item);
-  if(item.style.textDecoration == 'line-through')
-  item.style.textDecoration = 'none';
-  else
-  item.style.textDecoration = 'line-through';
-}
+  if (item.style.textDecoration == "line-through")
+    item.style.textDecoration = "none";
+  else item.style.textDecoration = "line-through";
+};
 
 const deleteTodo = (handleEvent) => {
-
   let item = handleEvent.target.parentNode;
-  if(document.querySelector('.todo-item').contains(".todo-item"))
-     document.querySelector('.todo-item').remove();
-}
+  item.addEventListener('transitionend', function(){
+    item.remove();
+  });
+  item.classList.add("to-do-list-item-fall");
+};
 
 todoTask.addEventListener("submit", (e) => {
   e.preventDefault();
   let input = document.querySelector("input");
-  if (input.value != "") 
-      getTodoList(input.value);
- input.value = "";
+  if (input.value != "") getTodoList(input.value);
+  input.value = "";
 });
 
-clearItems.addEventListener("click",(e) =>{
-    e.preventDefault();
-    let itemsPresent = document.querySelector(".to-do-list-item");
-    if(itemsPresent !== null)
-    itemsPresent.remove();
-})
+clearItems.addEventListener("click", (e) => {
+  e.preventDefault();
+  document.querySelector("ul").innerHTML = "";
+  // let itemsPresent = document.querySelector(".to-do-list-item");
+  // if(itemsPresent !== null)
+  // itemsPresent.remove();
+});
 
 const getTodoList = (tasklist) => {
   let ul = document.querySelector("ul");
@@ -50,6 +47,6 @@ const getTodoList = (tasklist) => {
           <button name="checkButton"><i class="fas fa-check-square"></i></button>
           <button name="deleteButton"><i class="fas fa-trash"></i></button>
     `;
-    li.classList.add('to-do-list-item');
-    ul.appendChild(li);
+  li.classList.add("to-do-list-item");
+  ul.appendChild(li);
 };
